% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{footstrap}
\alias{footstrap}
\alias{footstrap-package}
\title{Package title.}
\description{
What the package does (one paragraph).
}
\details{
Multi-line package description.
}
\section{Section name}{

You can add multiple sections.
}

\author{
\strong{Maintainer}: Erwin Gielens \email{first.last@example.com}

}
