% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bootstrap.R
\name{get_bootstrap}
\alias{get_bootstrap}
\title{The title goes here.}
\usage{
get_bootstrap(data, statistic, replicates, formula)
}
\arguments{
\item{data}{The argument type. Additional description for the argument.}

\item{statistic}{The argument type. Additional description for the argument.}

\item{replicates}{The argument type. Additional description for the argument.}

\item{formula}{The argument type. Additional description for the argument.}
}
\value{
Description about the return type of the function. Similarly, the return 
description can also span over multiple rows and the spacing matters.
}
\description{
The subtitle goes here.
}
\details{
A long description that spans over multiple rows is allowed right here. 
Note that `roxygen2` documentation starts with `#'`. Also, spacing matters.
}
\examples{
\dontrun{
test <- get.bootstrap(data = cars, statistic = get.lm, replicates = 1e4, formula = speed ~ dist)
}
}
